# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BizCore Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: Get all categories for the business
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /inventory:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: Get inventory entries for a location
      parameters:
        - name: locationId
          in: query
          required: true
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Inventory entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryEntry"
  /inventory/transactions:
    get:
      operationId: getInventoryTransactions
      tags: [inventory]
      summary: Get inventory transaction history
      parameters:
        - name: locationId
          in: query
          schema:
            type: integer
        - name: variantId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Inventory transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryTransaction"
    post:
      operationId: createInventoryTransaction
      tags: [inventory]
      summary: Record an inventory transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryTransactionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryTransaction"
  /inventory/low-stock:
    get:
      operationId: getLowStockItems
      tags: [inventory]
      summary: Get items below their low stock threshold
      parameters:
        - name: locationId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Low stock items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryEntry"
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        name:
          type: string
        parentId:
          type: ["integer", "null"]
        sortOrder:
          type: ["integer", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessId
        - name
        - active
        - createdAt
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        parentId:
          type: ["integer", "null"]
        sortOrder:
          type: integer
      required:
        - name
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: string
        parentId:
          type: ["integer", "null"]
        sortOrder:
          type: integer
        active:
          type: boolean
...
    InventoryEntry:
      type: object
      properties:
        id:
          type: integer
        variantId:
          type: integer
        locationId:
          type: integer
        quantity:
          type: string
        lowStockThreshold:
          type: ["string", "null"]
        isLowStock:
          type: boolean
        itemId:
          type: integer
        itemName:
          type: string
        variantName:
          type: string
        sku:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
      required:
        - id
        - variantId
        - locationId
        - quantity
        - isLowStock
        - itemId
        - itemName
        - variantName
    InventoryTransaction:
      type: object
      properties:
        id:
          type: integer
        variantId:
          type: integer
        locationId:
          type: integer
        type:
          type: string
        quantityChange:
          type: string
        referenceType:
          type: ["string", "null"]
        referenceId:
          type: ["integer", "null"]
        batchId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        itemName:
          type: ["string", "null"]
        variantName:
          type: ["string", "null"]
      required:
        - id
        - variantId
        - locationId
        - type
        - quantityChange
        - createdAt
    CreateInventoryTransactionBody:
      type: object
      properties:
        variantId:
          type: integer
        locationId:
          type: integer
        type:
          type: string
          enum: [purchase, adjustment, waste, return, transfer, sale]
        quantityChange:
          type: string
        notes:
          type: ["string", "null"]
        batchId:
          type: ["string", "null"]
      required:
        - variantId
        - locationId
        - type
        - quantityChange
...